crlite.ct_fetch_utils and refactor certs_to_crlite
